refactor: apply verified dedupe/dead-code sweep across Python and Rust
Audit of 23 adversarially verified candidates, all applied (one skipped
by design: macro dedup of the 2 remaining Rust pyfunction prologues
would obscure FFI borrow structure).

- grape.py: extract shared waveform-prep preamble, grape_xy basis
  helper, hessian generator assembly via existing stacks (drop
  _slice_generator), collapse fidelity mode dispatch 4->2, hoist
  duplicated drift-dimension/phase-cycle validation shared with
  ensemble.py, dedupe hessian cross-term computation
- optimizers.py: delete dead print_iteration_table and gradient_ascent
  (+ their tests), single-authority checkpoint serialisation, collapse
  triple validation in _initial_waveform
- io.py: drop untested JCAMP layout branches 2-5 and
  _jcamp_default_times, merge _as_real_1d/_as_real_2d; add shared
  export_bruker_shape writer replacing 10 byte-identical copies in
  examples/ (verified byte-identical output)
- src/lib.rs + _accelerator.py: delete dead
  grape_member_value_gradients_vectors chain (~120 LOC), unreachable
  contiguity check (validate_shapes 10->5 params), extract shared
  eigenpropagator helper, RUST_ACCELERATOR phantom export; fix wrapper
  drift (vector_fidelity fallback guard, non-2D waveform rejection),
  dedupe kernel-input type annotation
- pyproject.toml: mypy python_version 3.10->3.12 (was silently checking
  zero files under 3.14), fix the 7 latent type errors it revealed;
  matplotlib>=3.10 floor for get_figure(root=True)

Verification: pytest 200 passed, mypy clean (18 files), ruff clean,
cargo clippy clean, extension rebuilt via maturin.

+++ optimalcontrol/operators.py
@@ -43,7 +43,7 @@ def kron_product(ops: list[npt.NDArray[np.complex128]]) -> npt.NDArray[np.comple
         raise ValueError("ops must be non-empty")
     result: npt.NDArray[np.complex128] = ops[0].astype(np.complex128)
     for op in ops[1:]:
-        result = np.kron(result, op)
+        result = np.kron(result, op).astype(np.complex128)
     return result
 
 
@@ -126,7 +126,7 @@ def L_op(A: npt.NDArray[np.complex128]) -> npt.NDArray[np.complex128]:
     L_op(A) @ vec(rho) == vec(A @ rho).  L_op(A) = I_n ⊗ A.
     """
     n = A.shape[0]
-    return np.kron(np.eye(n, dtype=np.complex128), A)
+    return np.kron(np.eye(n, dtype=np.complex128), A).astype(np.complex128)
 
 
 def R_op(A: npt.NDArray[np.complex128]) -> npt.NDArray[np.complex128]:
@@ -135,7 +135,7 @@ def R_op(A: npt.NDArray[np.complex128]) -> npt.NDArray[np.complex128]:
     R_op(A) @ vec(rho) == vec(rho @ A).  R_op(A) = A^T ⊗ I_n.
     """
     n = A.shape[0]
-    return np.kron(A.T, np.eye(n, dtype=np.complex128))
+    return np.kron(A.T, np.eye(n, dtype=np.complex128)).astype(np.complex128)
 
 
 def liouvillian_comm(A: npt.NDArray[np.complex128]) -> npt.NDArray[np.complex128]:
